+++ RELEASE_NOTES.md
@@ -1,5 +1,10 @@
-# Release Notes — v0.3.0 (2026-06-09)
+# Release Notes — v0.3.2 (2026-06-09)
 
-### Fixed
+### Added
 
-- pipx installation issues
+- docs-website
+- `node_modules/`, `dist/` and `build/` at installed .gitignore
+
+### Changed
+
+- moved ai-architecture-cookbook MCP server from local install to npx repo